% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplot.R
\name{buildplot}
\alias{buildplot}
\title{Builds plotly interactive plot for each diagnostic function}
\usage{
buildplot(plotdf, ptitle, pxlab, pylab, hlinew = F, makeline = F)
}
\arguments{
\item{plotdf}{dataframe with x, y, and isprob data}

\item{pxlab}{x label}

\item{pylab}{y label}

\item{hlinew}{whether or not to make horizontal line at y = 0}

\item{makeline}{whether or not to fit y = x abline}

\item{ptite}{title of plot}
}
\value{
g Interactive plot of class plotly
}
\description{
Builds plotly interactive plot for each diagnostic function
}
