% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_reg.R
\name{int_reg}
\alias{int_reg}
\title{Creates an interactive regression plot for a given regression model.}
\usage{
int_reg(mod)
}
\arguments{
\item{mod}{The model that needs to be plotted}
}
\value{
An interactive plot or plots of the model
}
\description{
Creates an interactive regression plot for a given regression model.
}
