% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levplot.R
\name{levplot}
\alias{levplot}
\title{Produces an interactive leverage plot from an lm object}
\usage{
levplot(mod)
}
\arguments{
\item{mod}{an object of class lm}
}
\value{
g Interactive leverage plot of class Plotly
}
\description{
Produces an interactive leverage plot from an lm object
}
