% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.R
\name{MC}
\alias{MC}
\title{Calculates VIF and correlation matrix for a given Multiple Regression equation.}
\usage{
MC(dataset, eq)
}
\arguments{
\item{dataset}{The dataset to be used}

\item{equation}{A multiple regression equation supplied by user}
}
\value{
Plot and VIF
}
\description{
Calculates VIF and correlation matrix for a given Multiple Regression equation.
}
