% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.R
\name{MR}
\alias{MR}
\title{Calculates VIF and interactive correlation matrix for a given Multiple Regression equation.}
\usage{
MR(eq)
}
\arguments{
\item{eq}{A multiple regression equation supplied by user}
}
\value{
Plot and VIF
}
\description{
Calculates VIF and interactive correlation matrix for a given Multiple Regression equation.
}
