% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_reg.R
\name{multi_reg}
\alias{multi_reg}
\title{Plots the interactive simple regression or multiple regression with no predicted x}
\usage{
multi_reg(mod)
}
\arguments{
\item{mod}{The regression model}
}
\value{
An interactive plot or plots
}
\description{
Plots the interactive simple regression or multiple regression with no predicted x
}
