% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.R
\name{qq}
\alias{qq}
\title{Produces an interactive leverage plot from an lm object}
\usage{
qq(mod)
}
\arguments{
\item{mod}{an object of class lm}
}
\value{
g Interactive qqnorm plot of class Plotly
}
\description{
Produces an interactive leverage plot from an lm object
}
