% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplot.R
\name{resplot}
\alias{resplot}
\title{Produces an interactive residual plot from an lm object}
\usage{
resplot(mod)
}
\arguments{
\item{mod}{an object of class lm}
}
\value{
g Interactive residual plot of class plotly
}
\description{
Produces an interactive residual plot from an lm object
}
