% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIF_table.R
\name{VIF_table}
\alias{VIF_table}
\title{Calculates VIF for given multiple regression model}
\usage{
VIF_table(eq, limit)
}
\arguments{
\item{eq}{A multiple regression equation supplied by user
.@param limit A user defined limit for VIF value}
}
\value{
VIF with decision if the value is too high or not
}
\description{
Calculates VIF for given multiple regression model
}
